#24: re-symlinked all the non-changed files from CUO base
rolling forward to 2/27/23 CUO commit 9deadae5 - CUO version 0.1.11.0 incremental to 1.0.0.0
Split ClassicUO.Assets into two libraries

IO is for reading from disk. Assets is for managing in-memory assets.
The lines between these libraries are not at all right, but this was the
simplest way to separate the files for now.